# apps/runwisp/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
            "type": "string"
          },
          "message": {
            "description": "Error message text",
            "type": "string"
          },
          "value": {
            "description": "The value at the given location"
          }
        },
        "type": "object"
      },
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "http://localhost:9477/schemas/ErrorModel.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "LaunchTicketBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "http://localhost:9477/schemas/LaunchTicketBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "ticket": {
            "description": "Single-use, short-TTL ticket. Redeem via GET /api/auth/launch-ticket?ticket=...",
            "type": "string"
          }
        },
        "required": [
          "ticket"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "RunWisp API",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/launch-ticket": {
      "get": {
        "description": "Consumes the ticket, sets the session cookie, and redirects (303) to the `redirect` query parameter (a same-origin path; defaults to \"/\", and falls back to \"/\" if the value is unsafe).",
        "operationId": "redeemLaunchTicket",
        "parameters": [
          {
            "description": "Single-use ticket minted by POST /api/auth/launch-ticket",
            "explode": false,
            "in": "query",
            "name": "ticket",
            "required": true,
            "schema": {
              "description": "Single-use ticket minted by POST /api/auth/launch-ticket",
              "type": "string"
            }
          },
          {
            "description": "Same-origin absolute path to land on after the session cookie is set; defaults to /",
            "explode": false,
            "in": "query",
            "name": "redirect",
            "schema": {
              "description": "Same-origin absolute path to land on after the session cookie is set; defaults to /",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Redeem a single-use launch ticket and start a browser session",
        "tags": [
          "Auth"
        ]
      },
      "post": {
        "description": "The ticket can be redeemed via GET /api/auth/launch-ticket?ticket=\u003cticket\u003e.",
        "operationId": "createLaunchTicket",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaunchTicketBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Mint a single-use launch ticket",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:9477"
    }
  ]
}